NewVersion/Release.0.2.0

+++ src/main/java/com/packsendme/lib/tolls/response/dto/TollsCountryResponse_Dto.java
@@ -17,11 +17,11 @@ public class TollsCountryResponse_Dto implements Serializable {
 	public String name_country;
 	public int toll_amount;
 	public Double average_price_toll;
-	public int distance_country;
+	public Double distance_country;
 	public String currency_price;
 	
 	public TollsCountryResponse_Dto(String name_country, int toll_amount, Double average_price_toll,
-			int distance_country, String currency_price) {
+			Double distance_country, String currency_price) {
 		super();
 		this.name_country = name_country;
 		this.toll_amount = toll_amount;
@@ -29,6 +29,7 @@ public class TollsCountryResponse_Dto implements Serializable {
 		this.distance_country = distance_country;
 		this.currency_price = currency_price;
 	}
+
